+++ src/main/java/edu/illinois/library/cantaloupe/status/HealthChecker.java
@@ -4,31 +4,25 @@ import edu.illinois.library.cantaloupe.async.ThreadPool;
 import edu.illinois.library.cantaloupe.cache.CacheFacade;
 import edu.illinois.library.cantaloupe.cache.DerivativeCache;
 import edu.illinois.library.cantaloupe.cache.SourceCache;
-import edu.illinois.library.cantaloupe.image.Format;
 import edu.illinois.library.cantaloupe.image.Identifier;
 import edu.illinois.library.cantaloupe.image.Info;
-import edu.illinois.library.cantaloupe.operation.OperationList;
-import edu.illinois.library.cantaloupe.processor.Processor;
-import edu.illinois.library.cantaloupe.processor.ProcessorConnector;
-import edu.illinois.library.cantaloupe.processor.ProcessorFactory;
-import edu.illinois.library.cantaloupe.processor.SourceFormatException;
 import edu.illinois.library.cantaloupe.source.Source;
 import edu.illinois.library.cantaloupe.util.Stopwatch;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.imageio.stream.ImageInputStream;
+import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.HashSet;
-import java.util.Iterator;
 import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 
 /**
@@ -36,12 +30,10 @@ import java.util.concurrent.TimeUnit;
  * functioning correctly:</p>
  *
  * <dl>
- *     <dt>Processing I/O</dt>
+ *     <dt>Source I/O</dt>
  *     <dd>When an image endpoint successfully completes a request, it calls
- *     {@link #addSourceProcessorPair(Source, Processor, OperationList)} to
- *     register the various objects it used to do so. This class keeps track of
- *     every unique source-processor pair and tests each one against the last
- *     known image it worked with.</dd>
+ *     {@link #addSourceUsage(Source)} to register the source it used to do so.
+ *     This class tests the I/O of each unique source.</dd>
  *     <dt>The source cache</dt>
  *     <dd>An image is written to the source cache (if available) and read
  *     back.</dd>
@@ -58,127 +50,81 @@ public final class HealthChecker {
     private static final Logger LOGGER =
             LoggerFactory.getLogger(HealthChecker.class);
 
-    private static final Set<SourceProcessorPair> SOURCE_PROCESSOR_PAIRS =
+    private static final Set<SourceUsage> SOURCE_USAGES =
             ConcurrentHashMap.newKeySet();
 
     /**
-     * Can be set during testing to cause {@link #check()} to return a custom
-     * instance.
+     * Can be set during testing to cause {@link #checkSerially()} to return a
+     * custom instance.
      */
     private static Health overriddenHealth;
 
     /**
-     * <p>Informs the class of a {@link Source}-{@link Processor} pair that has
-     * been used successfully, and could be used again in the course of a
-     * health check. Should be called by image processing endpoints after
-     * processing has completed successfully.</p>
+     * <p>Informs the class of a {@link Source} that has just been used
+     * successfully, and could be used again in the course of a health check.
+     * Should be called by image processing endpoints after processing has
+     * completed successfully.</p>
      *
      * <p>This method is thread-safe.</p>
      */
-    public static void addSourceProcessorPair(Source source,
-                                              Processor processor,
-                                              OperationList opList) {
-        final SourceProcessorPair pair = new SourceProcessorPair(
-                source, processor.getClass().getName(), opList);
+    public static void addSourceUsage(Source source) {
+        final SourceUsage usage = new SourceUsage(source);
         // The pair is configured to read a specific image. We want to remove
         // any older "equal" (see this ivar's equals() method!) instance before
         // adding the current one because the older one is more likely to be
         // stale (no longer accessible), in light of the possibility that the
         // application has been running for a while.
-        SOURCE_PROCESSOR_PAIRS.remove(pair);
-        SOURCE_PROCESSOR_PAIRS.add(pair);
+        SOURCE_USAGES.remove(usage);
+        SOURCE_USAGES.add(usage);
     }
 
     /**
      * For testing only!
      */
-    public static Set<SourceProcessorPair> getSourceProcessorPairs() {
-        return SOURCE_PROCESSOR_PAIRS;
+    public static Set<SourceUsage> getSourceUsages() {
+        return SOURCE_USAGES;
     }
 
     /**
      * For testing only!
      *
-     * @param health Custom instance that will be returned by {@link #check()}.
-     *               Supply {@code null} to clear the override.
+     * @param health Custom instance that will be returned by {@link
+     *               #checkSerially()}. Supply {@code null} to clear the
+     *               override.
      */
     public static synchronized void overrideHealth(Health health) {
         overriddenHealth = health;
     }
 
     /**
-     * <p>Checks the functionality of every {@link #addSourceProcessorPair(
-     * Source, Processor, OperationList) known} source-processor pair. The
-     * checks exercise the full length of the processing pipeline, reading an
-     * image from a {@link Source}, running it through a {@link Processor}, and
-     * writing it to an output stream.</p>
+     * <p>Checks the functionality of every {@link #addSourceUsage(Source)
+     * known source}.</p>
      *
      * <p>The individual checks are done concurrently in as many threads as
-     * there are unique pairs. This is intended to improve responsiveness,
-     * assuming there aren't a whole lot more pairs than there are CPU
-     * threads.</p>
+     * there are unique pairs.</p>
      */
-    private static synchronized void checkProcessing(Health health) {
+    private static synchronized void checkSources(Health health) {
         // Make a local copy to ensure that another thread doesn't change it
         // underneath us.
-        final Set<SourceProcessorPair> localPairs =
-                new HashSet<>(SOURCE_PROCESSOR_PAIRS);
-        final int numPairs = localPairs.size();
+        final Set<SourceUsage> localUsages = new HashSet<>(SOURCE_USAGES);
+        final int numUsages                = localUsages.size();
 
-        LOGGER.trace("{} unique source/processor combinations.", numPairs);
+        LOGGER.trace("{} unique sources.", numUsages);
 
-        final CountDownLatch latch = new CountDownLatch(numPairs);
-        localPairs.forEach(pair -> {
+        final CountDownLatch latch = new CountDownLatch(numUsages);
+        localUsages.forEach(usage -> {
             ThreadPool.getInstance().submit(() -> {
-                LOGGER.debug("Exercising processing I/O: {}", pair);
-
-                Future<Path> tempFileFuture = null;
-                final ProcessorFactory pf = new ProcessorFactory();
-                final Source source = pair.getSource();
-
-                final Iterator<Format> formatIterator = source.getFormatIterator();
-
-                // This should never be the case, but is a basic sanity check.
-                if (!formatIterator.hasNext()) {
+                final Source source = usage.getSource();
+                LOGGER.trace("Exercising source I/O for {}", usage);
+                // TODO: it would be nice to have a Source.exists() method here
+                try (ImageInputStream is = source.newStreamFactory().newSeekableStream()) {
+                    is.length();
+                } catch (IOException e) {
                     health.setMinColor(Health.Color.RED);
-                    health.setMessage("Format iterator returned an empty " +
-                            "iterator. This is probably a bug.");
+                    health.setMessage(String.format("%s (%s)",
+                            e.getMessage(), usage));
+                } finally {
                     latch.countDown();
-                }
-
-                while (formatIterator.hasNext()) {
-                    final Format format = formatIterator.next();
-                    try (Processor processor = pf.newProcessor(pair.getProcessorName());
-                         OutputStream os = OutputStream.nullOutputStream()) {
-                        processor.setSourceFormat(format);
-
-                        tempFileFuture = new ProcessorConnector().connect(
-                                source,
-                                processor,
-                                source.getIdentifier(),
-                                format);
-                        Info info = processor.readInfo();
-                        processor.process(pair.getOperationList(), info, os);
-                    } catch (SourceFormatException e) {
-                        LOGGER.debug("checkProcessing(): {}", e.getMessage());
-                    } catch (Throwable t) {
-                        health.setMinColor(Health.Color.RED);
-                        health.setMessage(String.format("%s (%s)",
-                                t.getMessage(), pair));
-                    } finally {
-                        latch.countDown();
-                        if (tempFileFuture != null) {
-                            try {
-                                Path tempFile = tempFileFuture.get();
-                                if (tempFile != null) {
-                                    Files.deleteIfExists(tempFile);
-                                }
-                            } catch (Exception e) {
-                                LOGGER.error("checkProcessing(): failed to delete temp file: {}",
-                                        e.getMessage(), e);
-                            }
-                        }
-                    }
                 }
             });
         });
@@ -199,12 +145,11 @@ public final class HealthChecker {
         final Optional<SourceCache> optSrcCache = cacheFacade.getSourceCache();
         if (optSrcCache.isPresent()) {
             final SourceCache srcCache = optSrcCache.get();
-            LOGGER.debug("Exercising the source cache: {}", srcCache);
+            LOGGER.trace("Exercising the source cache: {}", srcCache);
             final Identifier identifier =
                     new Identifier("HealthCheck-" + UUID.randomUUID());
             try {
                 // Exercise the cache. Errors will throw exceptions.
-                srcCache.purge(identifier);
                 // Write a file to disk.
                 try (OutputStream os = srcCache.newSourceImageOutputStream(identifier)) {
                     String message = String.format("This file written by %s",
@@ -216,6 +161,8 @@ public final class HealthChecker {
                 // Read it back.
                 Optional<Path> path = srcCache.getSourceImageFile(identifier);
                 Files.readAllBytes(path.orElseThrow());
+                // Delete it.
+                srcCache.purge(identifier);
             } catch (Throwable t) {
                 health.setMinColor(Health.Color.RED);
                 String message = String.format("%s: %s",
@@ -235,16 +182,17 @@ public final class HealthChecker {
                 cacheFacade.getDerivativeCache();
         if (optDerivativeCache.isPresent()) {
             DerivativeCache dCache = optDerivativeCache.get();
-            LOGGER.debug("Exercising the derivative cache: {}", dCache);
+            LOGGER.trace("Exercising the derivative cache: {}", dCache);
             final Identifier identifier =
                     new Identifier("HealthCheck-" + UUID.randomUUID());
             try {
                 // Exercise the cache. Errors will throw exceptions.
-                dCache.purge(identifier);
                 // Write to the cache.
                 dCache.put(identifier, new Info());
                 // Read it back.
                 dCache.getInfo(identifier);
+                // Delete it.
+                dCache.purge(identifier);
             } catch (Throwable t) {
                 health.setMinColor(Health.Color.RED);
                 String message = String.format("%s: %s",
@@ -257,22 +205,18 @@ public final class HealthChecker {
 
     /**
      * <p>Performs a health check as explained in the class documentation.
-     * Each group of checks (derivative cache, processor I/O, etc.) is
-     * performed sequentially. If any check fails, all remaining checks are
-     * skipped.</p>
+     * Each group of checks (derivative cache, source I/O, etc.) is performed
+     * sequentially. If any check fails, all remaining checks are skipped.</p>
      *
-     * <p>N.B.: there could be benefits in terms of responsiveness to doing all
-     * of the checks asynchronously and returning a result when they've all
-     * completed; but this could also result in having to do more checks than
-     * necessary, since, if a single check fails, none of the others
-     * matter.</p>
-     *
-     * <p>This method is thread-safe and can be called repeatedly to obtain the
-     * current health.</p>
+     * <p>N.B.: it may be faster to {@link #checkConcurrently() do all of the
+     * checks concurrently}; but this could also result in having to do more
+     * checks than necessary, since, if a single check fails, none of the
+     * others matter.</p>
      *
      * @return Instance reflecting the application health.
+     * @see #checkConcurrently()
      */
-    public Health check() {
+    public Health checkSerially() {
         LOGGER.debug("Initiating a health check");
 
         if (overriddenHealth != null) {
@@ -282,10 +226,10 @@ public final class HealthChecker {
         final Stopwatch watch = new Stopwatch();
         final Health health   = new Health();
 
-        // Check processing I/O.
+        // Check source input.
         if (!Health.Color.RED.equals(health.getColor())) {
-            checkProcessing(health);
-            LOGGER.trace("Processing I/O check completed in {}; health so far is {}",
+            checkSources(health);
+            LOGGER.trace("Source I/O check completed in {}; health so far is {}",
                     watch, health.getColor());
         }
 
@@ -305,9 +249,82 @@ public final class HealthChecker {
 
         // Log the final status.
         if (Health.Color.GREEN.equals(health.getColor())) {
-            LOGGER.debug("Health check completed in {}: {}", watch, health);
+            LOGGER.debug("Sequential health check completed in {}: {}",
+                    watch, health);
         } else {
-            LOGGER.warn("Health check completed in {}: {}", watch, health);
+            LOGGER.warn("Sequential health check completed in {}: {}",
+                    watch, health);
+        }
+        return health;
+    }
+
+    /**
+     * <p>Performs a health check as explained in the class documentation.
+     * Each group of checks (derivative cache, source I/O, etc.) is performed
+     * concurrently. This may cause the check to complete faster than the one
+     * via {@link #checkSerially()} but it also may result in performing more
+     * checks than necessary.</p>
+     *
+     * @return Instance reflecting the application health.
+     * @see #checkSerially()
+     */
+    public Health checkConcurrently() {
+        LOGGER.debug("Initiating a health check");
+
+        if (overriddenHealth != null) {
+            return overriddenHealth;
+        }
+
+        final Stopwatch watch      = new Stopwatch();
+        final Health health        = new Health();
+        final ThreadPool pool      = ThreadPool.getInstance();
+        final CountDownLatch latch = new CountDownLatch(3);
+
+        // Check source I/O.
+        pool.submit(() -> {
+            try {
+                checkSources(health);
+                LOGGER.trace("Source I/O check completed in {}; health so far is {}",
+                        watch, health.getColor());
+            } finally {
+                latch.countDown();
+            }
+        });
+
+        // Check the source cache.
+        pool.submit(() -> {
+            try {
+                checkSourceCache(health);
+                LOGGER.trace("Source cache check completed in {}; health so far is {}",
+                        watch, health.getColor());
+            } finally {
+                latch.countDown();
+            }
+        });
+
+        // Check the derivative cache.
+        pool.submit(() -> {
+            try {
+                checkDerivativeCache(health);
+                LOGGER.trace("Derivative cache check completed in {}; health so far is {}",
+                        watch, health.getColor());
+            } finally {
+                latch.countDown();
+            }
+        });
+
+        try {
+            latch.await(60, TimeUnit.SECONDS);
+        } catch (InterruptedException ignore) {
+        }
+
+        // Log the final status.
+        if (Health.Color.GREEN.equals(health.getColor())) {
+            LOGGER.debug("Concurrent health check completed in {}: {}",
+                    watch, health);
+        } else {
+            LOGGER.warn("Concurrent health check completed in {}: {}",
+                    watch, health);
         }
         return health;
     }